+++ backend/src/Controller/ServiceController.php
@@ -10,9 +10,16 @@ use App\Entity\Image;
 use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
-
+use App\Service\JwtService;
 final class ServiceController extends AbstractController
 {
+    private JwtService $jwtService;
+
+    public function __construct(JwtService $jwtService)
+    {
+        $this->jwtService = $jwtService;
+    }
+
     // GET ALL SERVICES
     #[Route('/services', name: 'get_services', methods: ['GET'])]
     public function getServices(EntityManagerInterface $entityManager): JsonResponse
@@ -34,6 +41,13 @@ final class ServiceController extends AbstractController
     public function createService(EntityManagerInterface $entityManager, Request $request): JsonResponse
     {
         $data = json_decode($request->getContent(), true);
+
+        // CHECK IF USER IS ADMIN OR EMPLOYEE
+        $isAdmin = $this->jwtService->checkEmployee($request->headers->get('Authorization'), $entityManager);
+        if (!$isAdmin) {
+            return $this->json(['message' => 'Unauthorized'], Response::HTTP_UNAUTHORIZED);
+        }
+
         $service = new Service();
         $service->setName($data['name']);
         $service->setDescription($data['description']);
@@ -57,6 +71,13 @@ final class ServiceController extends AbstractController
     public function updateService(EntityManagerInterface $entityManager, Request $request, int $id): JsonResponse
     {
         $data = json_decode($request->getContent(), true);
+
+        // CHECK IF USER IS ADMIN OR EMPLOYEE
+        $isAdmin = $this->jwtService->checkEmployee($request->headers->get('Authorization'), $entityManager);
+        if (!$isAdmin) {
+            return $this->json(['message' => 'Unauthorized'], Response::HTTP_UNAUTHORIZED);
+        }
+
         $service = $entityManager->getRepository(Service::class)->find($id);
         $service->setName($data['name']);
         $service->setDescription($data['description']);
@@ -78,8 +99,14 @@ final class ServiceController extends AbstractController
 
     // DELETE SERVICE
     #[Route('/services/{id}', name: 'delete_service', methods: ['DELETE'])]
-    public function deleteService(EntityManagerInterface $entityManager, int $id): JsonResponse
+    public function deleteService(Request $request, EntityManagerInterface $entityManager, int $id): JsonResponse
     {
+        // CHECK IF USER IS ADMIN OR EMPLOYEE
+        $isAdmin = $this->jwtService->checkEmployee($request->headers->get('Authorization'), $entityManager);
+        if (!$isAdmin) {
+            return $this->json(['message' => 'Unauthorized'], Response::HTTP_UNAUTHORIZED);
+        }
+
         $service = $entityManager->getRepository(Service::class)->find($id);
         $entityManager->remove($service);
         $entityManager->flush();
